extract Event to its own component
